✨ feat: introduce caching for embeddings creation
Added caching logic for embeddings via createCachedEmbedder function.
